Make Statistics directly usable as Metrics, by having them implement the Metric interface by passing through the mean of the statistic.
This allows a lot of boilerplate Metrics to be removed which were trivially wrapping a Statistic; in these cases a Metric of the same name is still exposed by renaming or aliasing the Statistic, which is now itself a Metric.

NoOpMetric (which was a general way to wrap a statistic as a metric) is also no longer needed, however a backwards-compatibility shim for it remains.

PiperOrigin-RevId: 778907738