Implement BCa (bias-corrected and accelerated) bootstrap confidence intervals.
BCa intervals use the jackknife (a leave-one-out method) to estimate an 'acceleration' parameter (related to skewness) that it uses to correct the CI quantiles.
For the IID bootstrap we can use the standard jackknife for this as in the original paper.
For the stationary bootstrap (with temporally dependent data) we use the stationary jackknife, which is a leave-block-out method similar to the stationary bootstrap.

Also adds more detail and citations to statistical_inference/README.md.

PiperOrigin-RevId: 819726355