Add a helper to compute per-unit values of the metric after linearizing it around the mean statistics. This uses JAX's autodiff.
The mean of these per-unit values is equal to the value of the metric, and a variance estimate for the mean of the linearized values approximates the variance of the value of the metric, as explained in docstrings.

This will allow statistical methods developed for inference of simple means, to be used in the more general setting of a non-linear function of means.

PiperOrigin-RevId: 822952507