Standardize on UnbiasedSpreadSkillRatio, removing SpreadSkillRatio which lacks the desired property that numerator and denominator are equal in expectation for perfect forecasts, except in the limit of large ensembles.
While it would be possible to correct the SpreadSkillRatio as implemented using a correction factor described in [1], using the unbiased MSE estimator for skill is a simpler way for us to achieve the desired property. It is also somewhat redundant to have two spread-skill ratios implemented as they are deterministic functions of eachother (at least for constant ensemble size).

Also fixes the test of the spread-skill ratio, which was using a tolerance that was way too loose (atol of about 0.9 for a test comparing with 1).

[1] Fortin et al, Why Should Ensemble Spread Match the RMSE of the Ensemble Mean? J. Hydrometeorol. 15, 1708–1713 (2014).

PiperOrigin-RevId: 830898121