Add support for a version of the t-test based on a HAC (Heteroskedasticity and Autocorrelation Consistent) estimator with default settings following the recommendations of Lazarus et al [1].
Their recommendation is to use an EWC estimator, with a specific carefully-optimized scaling rule and constant factor for the spectral cut-off used by this estimator. The method is related to the Diebold-Mariano test when used to compare forecast skill of two models.

As part of this, refactors t_test.py into an abstract base class with 3 implementations: t_test.IID, t_test.GeerAR2Corrected and t_test.LazarusHACEWC.

[1] Lazarus, E., Lewis, D. J., Stock, J. H. & Watson, M. W. HAR inference: Recommendations for practice. J. Bus. Econ. Stat. 36, 541–559 (2018).

PiperOrigin-RevId: 831450751